# openapi.json
{
...
  "info": {
    "description": "Auto-generated route inventory for this iterion instance. Routes and methods are exact; request/response schemas are populated for the typed surface (see components) and enriched route-by-route. Note: the /api/v1/native, /api/v1/dispatcher and /api/v1/mcp/board sub-trees are served but registered on a separate mux, so they are intentionally not catalogued here.",
    "title": "iterion API",
    "version": "current"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/examples/{name}": {
      "get": {
        "operationId": "getExamplesByName",
        "responses": {
          "default": {
            "description": "Response"
          }
        },
        "summary": "GET /api/examples/{name...}",
        "tags": [
          "examples"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "name",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
    "/api/runs/{id}/files": {
      "get": {
        "operationId": "getRunsByIdFiles",
        "responses": {
          "default": {
            "description": "Response"
          }
        },
        "summary": "GET /api/runs/{id}/files",
        "tags": [
          "runs"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
    "/api/v1/bots/{name}": {
      "get": {
        "operationId": "getV1BotsByName",
        "responses": {
          "default": {
            "description": "Response"
          }
        },
        "summary": "GET /api/v1/bots/{name}",
        "tags": [
          "bots"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "name",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "put": {
        "operationId": "putV1BotsByName",
        "responses": {
          "default": {
            "description": "Response"
          }
        },
        "summary": "PUT /api/v1/bots/{name}",
        "tags": [
          "bots"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}